# contracts/openapi-v1.yaml
openapi: 3.0.3
info:
  title: Data Structure Agent API
  version: 1.0.0
  description: Spring backend contract for the data structure learning agent.
servers:
  - url: /api/v1
    description: Relative base; usable from structify.cn and admin.structify.cn. Admin routes still require a database-backed ADMIN role.
security: []
...
paths:
...
  /chat:
    post:
      tags: [Chat]
      summary: Complete a course-grounded chat turn
      security: [{cookieAuth: []}, {bearerAuth: []}]
      description: Formal course-grounded generation requires an authenticated active user, published authorized evidence, and an enabled model with an available daily token quota. The request is metered through an idempotent reservation and settled with provider usage when available, otherwise the reservation estimate.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChatRequest'}
      responses:
        '200':
          description: Assistant answer and reviewed sources
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ChatResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '502': {$ref: '#/components/responses/BadGateway'}
        '504': {$ref: '#/components/responses/GatewayTimeout'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /chat/stream:
    post:
      tags: [Chat]
      summary: Stream a course-grounded chat turn
      security: [{cookieAuth: []}, {bearerAuth: []}]
      description: Formal generation requires an authenticated active user and the same published-evidence and quota gate as POST /chat. After the request is accepted, the response is an SSE stream with named sources, delta, done, and error events. The server releases the reservation and concurrency slot when the upstream fails, the stream times out, or the client disconnects.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChatRequest'}
      responses:
        '200':
          description: Server-sent event stream
          content:
            text/event-stream:
              schema: {type: string}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '502': {$ref: '#/components/responses/BadGateway'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
        '504': {$ref: '#/components/responses/GatewayTimeout'}
  /chat/sessions:
    get:
      tags: [Chat]
      summary: List the authenticated student's persisted chat sessions
      security: [{cookieAuth: []}, {bearerAuth: []}]
      responses:
        '200':
          description: The 50 most recently active sessions, ordered newest first
          content:
            application/json:
              schema:
                type: array
                maxItems: 50
                items: {$ref: '#/components/schemas/ChatSessionSummary'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /chat/sessions/{sessionId}:
    get:
      tags: [Chat]
      summary: Read one owned chat session and its messages
      security: [{cookieAuth: []}, {bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/ChatSessionId'
      responses:
        '200':
          description: Owned session with its latest 200 messages in chronological order
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ChatSession'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
    delete:
      tags: [Chat]
      summary: Delete one owned persisted chat session
      security: [{cookieAuth: []}, {bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/ChatSessionId'
      responses:
        '204': {description: Session deleted}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
...
  /learning/progress:
    get:
      tags: [Learning]
      summary: Aggregate owned learning activity by chapter
      security: [{cookieAuth: []}, {bearerAuth: []}]
      responses:
        '200':
          description: Explainable activity counts, not an opaque score
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LearningProgress'}
        '401': {$ref: '#/components/responses/Unauthorized'}
  /learning/events:
    post:
      tags: [Learning]
      summary: Record an authenticated student's reviewed learning activity
      security: [{cookieAuth: []}, {bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LearningEventRequest'}
      responses:
        '201':
          description: Learning activity recorded
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LearningEvent'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '413': {$ref: '#/components/responses/PayloadTooLarge'}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: ds_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ChatSessionId:
      name: sessionId
      in: path
      required: true
      schema: {type: string, maxLength: 64}
...
  responses:
    BadRequest:
      description: Invalid request
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    Unauthorized:
      description: Authentication required or credentials invalid
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
...
    NotFound:
      description: Resource not found or not visible to this user
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    Conflict:
      description: State conflict
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    PayloadTooLarge:
      description: Input exceeds configured limit
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    TooManyRequests:
      description: Rate limit exceeded
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    BadGateway:
      description: Model or sandbox upstream failed
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    GatewayTimeout:
      description: Model or sandbox upstream timed out
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
    ServiceUnavailable:
      description: A required local configuration or controlled dependency is unavailable
      content: {application/json: {schema: {$ref: '#/components/schemas/ApiError'}}}
...
  schemas:
    ApiError:
      type: object
      required: [code, message, requestId, details]
      properties:
        code: {type: string}
        message: {type: string}
        requestId: {type: string}
        details:
          type: array
          items: {type: string}
...
    ChatTurn:
      type: object
      required: [role, content]
      properties:
        role: {type: string, enum: [user, assistant]}
        content: {type: string, maxLength: 4000}
    ChatRequest:
      type: object
      required: [prompt]
      properties:
        prompt: {type: string, maxLength: 4000}
        chapterId: {type: string, maxLength: 64}
        sessionId: {type: string, maxLength: 64}
        history:
          type: array
          maxItems: 12
          items: {$ref: '#/components/schemas/ChatTurn'}
    ChatSource:
      type: object
      required: [id, chapterId, title, content, source, score]
      properties:
        id: {type: string}
        chapterId: {type: string}
        title: {type: string}
        content: {type: string, maxLength: 500}
        source: {type: string}
        pageLabel: {type: string, nullable: true}
        score: {type: number, format: double}
    ChatResponse:
      type: object
      required: [answer, sources, persisted]
      properties:
        answer: {type: string}
        sessionId: {type: string, nullable: true}
        sources:
          type: array
          items: {$ref: '#/components/schemas/ChatSource'}
        persisted: {type: boolean}
    ChatSessionSummary:
      type: object
      required: [id, title, updatedAt, messageCount]
      properties:
        id: {type: string}
        chapterId: {type: string, nullable: true}
        title: {type: string}
        updatedAt: {type: string, format: date-time}
        messageCount: {type: integer, format: int64}
    ChatMessage:
      type: object
      required: [id, role, content, sources, createdAt]
      properties:
        id: {type: integer, format: int64}
        role: {type: string, enum: [user, assistant]}
        content: {type: string}
        sources:
          type: array
          items: {$ref: '#/components/schemas/ChatSource'}
        createdAt: {type: string, format: date-time}
    ChatSession:
      type: object
      required: [id, title, updatedAt, messages]
      properties:
        id: {type: string}
        chapterId: {type: string, nullable: true}
        title: {type: string}
        updatedAt: {type: string, format: date-time}
        messages:
          type: array
          maxItems: 200
          description: Latest persisted messages, returned in chronological order
          items: {$ref: '#/components/schemas/ChatMessage'}
...
    ChapterProgress:
      type: object
      required:
        - chapterId
        - chapterNumber
        - title
        - chatCount
        - classroomCount
        - animationCount
        - codeRunCount
        - eventCount
        - totalActivities
      properties:
        chapterId: {type: string}
        chapterNumber: {type: integer}
        title: {type: string}
        chatCount: {type: integer, format: int64}
        classroomCount: {type: integer, format: int64}
        animationCount: {type: integer, format: int64}
        codeRunCount: {type: integer, format: int64}
        eventCount: {type: integer, format: int64}
        totalActivities: {type: integer, format: int64}
        lastActivityAt: {type: string, format: date-time, nullable: true}
    LearningProgress:
      type: object
      required: [totalActivities, chapters]
      properties:
        totalActivities: {type: integer, format: int64}
        chapters:
          type: array
          items: {$ref: '#/components/schemas/ChapterProgress'}
    LearningEventRequest:
      type: object
      required: [eventType]
      properties:
        eventType:
          type: string
          description: Classroom answers, animation observations, and code reviews are recorded only by their domain endpoints.
          enum: [RESOURCE_VIEW, RESOURCE_DOWNLOAD, REVIEW_COMPLETED, WEAKNESS_RECORDED]
        chapterId: {type: string, maxLength: 64, nullable: true}
        referenceId: {type: string, maxLength: 96, nullable: true}
        payload: {type: object, additionalProperties: true, nullable: true}
    LearningEvent:
      type: object
      required: [id, eventType, createdAt]
      properties:
        id: {type: integer, format: int64}
        eventType: {type: string}
        chapterId: {type: string, nullable: true}
        referenceId: {type: string, nullable: true}
        createdAt: {type: string, format: date-time}
...